Version 35 fixes a bug in scrdit.c where ^<PSHTAB x> would set the type of tab x instead of ^<POPTAB x> (which modifies the value of tab x). This bug had existed since ^<{PSHTAB,POPTAB}> were introduced.
In fndtn.qm, macro '8' (N8) outputs the local date (i.e. gives keyboard access
to pseudomacro 4011)

In word.qm, The ^N^A suite now uses ^<SUB> instead of the (soon to be withdrawn)
^NY tab subtraction pseudomacro.

Also in word.qm, fix a bug in the ^N^A suite which filled the R stack after 8
invocations. ^N^A now:
 - saves the value of tab 1
 - forces tab 1 to 1
 - restores tab 1 to its former value on successful completeion