Q version 56.37: q -v works as documented again; introduce q -vv
During the version 56 development cycle, the action of q -v was changed to
ignore FD-FDEVNULL commands. The original function of q -v was to only affect
behaviour when running in a pipe and this has been restored. For the (rarer)
case where one does want to disable FD, q -v -v has been implemeted. This
disables FD in any invocation of Q.

Updated:

 help/iq: Document -vv "option"

 man/man1/q.1: Document action of -v given twice

 q/q.c: Implement -vv

 q/q_version.h: 56.37